Q3G7HM8 split lib/scrubber.ts into a scrubber/ directory, one module per question
